requirements.txt
discord.py>=2.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
openai>=1.30.0
PyNaCl>=1.5.0
yt-dlp>=2024.0.0
discord-ext-voice-recv>=0.5.0
